refactor(mcp): consolidate test suite per LLM Council review
- Merged test_e2e.py into test_server.py
- Removed 12 duplicate tests (31 → 19 tests)
- Kept 4 genuinely unique tests:
  - test_rate_limiter_concurrent_access
  - test_server_has_resources_registered
  - test_known_issues_resource_returns_content
  - test_utf8_pipes_configured
- Removed unused mcp_env fixture from conftest.py
- All 19 tests passing, ruff clean
